(trends) Added hot trends component
