+++ shared/components/keyword_page.jsx
@@ -8,7 +8,7 @@ import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
 import NavBar from './nav_bar';
 import TweetList from './twitter/tweet_list';
-import LandingPage from './landing_page';
+import HotTrends from './hot_trends';
 import TreeMapView from './List/treemap_view';
 
 class KeywordPage extends Component {
@@ -67,7 +67,7 @@ class KeywordPage extends Component {
             </div>
 
             <div className="col-sm-6">
-              <LandingPage/>
+              <HotTrends />
             </div>
           </div>
