+++ shared/components/landing_page.jsx
@@ -1,57 +1,12 @@
 import React from 'react';
 import { Component } from 'react';
-import SearchBar from './search_bar';
+import HotTrends from './hot_trends';
 import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
 import {getHotTrends} from '../actions/hotTrends';
 
-class LandingPage extends Component {
-  componentWillMount() {
-    this.props.getHotTrends();
-  }
-
-  renderDBTrends() {
-    return this.props.hotTrends.randomTen.map( (item,index) => {
-      if(index < 6 ) {
-      return (
-        <div className="col-sm-4" key={item}>
-          <div className='card card-block card-success text-xs-center'> {item} </div>
-        </div>
-      );
-    }
-    });
-  }
-
-  renderTweets() {
-    return this.props.hotTrends.stocks.map( (item,index) => {
-      if(index < 6) {
-      return (
-        <div className="col-sm-4" key={item}>
-          <div className='card card-block card-info text-xs-center'> {item} </div>
-        </div>
-      );
-    }
-    });
-  }
-  renderKeywords() {
-    return this.props.hotTrends.keywords.map( (item,index) => {
-      if(index < 6) {
-      return (
-        <div className="col-sm-4" key={item}>
-          <div className='card card-block card-success text-xs-center'> {item} </div>
-        </div>
-      );
-    }
-    });
-  }
-
+export default class LandingPage extends Component {
   render() {
-    if(!this.props.hotTrends) {
-      return (
-        <div> Loading... </div>
-      );
-    }
-
     return (
     <div>
       <div className="about-landing">
@@ -61,26 +16,8 @@ class LandingPage extends Component {
           Culture drives business. Trendr's goal is simple: demonstrate this relationship by visualizing correlations in big data. By combining Google Search and finance data, Trendr uses the Alchemy API and Natural Language Processing to find correlations between popular Google Search terms and companies.
         </p>
       </div>
-      <div className="list spacer">
-        <h2> Trends Last Month </h2>
-        <div className="row">
-          {this.renderDBTrends()}
-        </div>
-        <h2> Twitter Today </h2>
-        <div className="row">
-          {this.renderTweets()}
-        </div>
-        <h2> Trends This Hour </h2>
-        <div className="row">
-          {this.renderKeywords()}
-        </div>
-      </div>
+      <HotTrends />
     </div>
     );
   }
-}
-
-function mapStatesToProps(state) {
-  return {hotTrends: state.hotTrends.items};
-}
-export default connect(mapStatesToProps, {getHotTrends})(LandingPage);
+};
